Migrate methods to repositories
