+++ src/domain/instance/InstanceRepository.ts
@@ -2,4 +2,5 @@ import Instance from "./Instance";
 
 export default interface InstanceRepository {
     getById(id: string): Promise<Instance>;
+    getDefaultIds(filter?: string): Promise<string[]>;
 }
